requirements.txt
# Core
python-dotenv>=1.0.0

# Vietnamese NLP
underthesea>=6.8.0

# Embeddings & Reranking
sentence-transformers>=2.7.0
FlagEmbedding>=1.2.0

# Vector DB
qdrant-client>=1.9.0

# BM25
rank-bm25>=0.2.2

# Evaluation
ragas>=0.2.0
datasets>=2.0.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0

# LLM
openai>=1.0.0
requests>=2.31.0

# PDF processing
pymupdf>=1.24.0

# Testing & Linting
pytest>=8.0.0
ruff>=0.4.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
pyyaml>=6.0.0

# Guardrails
presidio-analyzer>=2.2.354
presidio-anonymizer>=2.2.354
transformers>=4.43.0
